+++ project.sh
@@ -31,5 +31,12 @@ $VENV/bin/redup scan . --format toon --output ./project
 #$VENV/bin/redup scan . --functions-only -f toon --output ./project
 #$VENV/bin/vallm batch ./src --recursive --semantic --model qwen2.5-coder:7b
 #$VENV/bin/vallm batch --parallel .
-$VENV/bin/vallm batch . --recursive --format toon --output ./project
-$VENV/bin/prefact -a -e "examples/**"
+#$VENV/bin/vallm batch . --recursive --format toon --output ./project
+#$VENV/bin/prefact -a -e "examples/**"
+
+$PIP install doql --upgrade --quiet
+$VENV/bin/doql adopt . --format less --force
+
+$PIP install sumd --upgrade --quiet
+$VENV/bin/sumd .
+$VENV/bin/sumr .